[#40] 결제 공통부분 템플릿-콜백 패턴 적용
- 결제 방법 선택 팩토리 -> payForOrder 메서드
- 변하지 않는 결제 정보, 재고차감로직 템플릿 화
- 변하는 부분은 추상화를 통한 전략패턴 적용
- 람다를 통한 콜백 메서드
