+++ src/main/java/me/jjeda/mall/orders/dto/CreditPaymentDto.java
@@ -1,4 +1,13 @@
 package me.jjeda.mall.orders.dto;
 
+import me.jjeda.mall.orders.domain.CreditPayment;
+
 public class CreditPaymentDto extends PaymentDto {
+    public static PaymentDto toDto(CreditPayment creditPayment) {
+        return null;
+    }
+
+    public CreditPayment toEntity() {
+        return null;
+    }
 }
