+++ src/main/java/me/jjeda/mall/orders/dto/MobilePaymentDto.java
@@ -1,4 +1,13 @@
 package me.jjeda.mall.orders.dto;
 
+import me.jjeda.mall.orders.domain.MobilePayment;
+
 public class MobilePaymentDto extends PaymentDto {
+    public static PaymentDto toDto(MobilePayment mobilePayment) {
+        return null;
+    }
+
+    public MobilePayment toEntity() {
+        return null;
+    }
 }
